scraper/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
supabase==2.3.4
googlemaps==4.10.0
black==24.2.0
flake8==7.0.0
